feat: add route navigation
The navigation between the different app components is performed by
using route navigation.

In addition to this, two new components have been added:

* RecipeStartComponent
* RecipeEditComponent

This first one shows a message to the user about selecting a recipe.
The second one will be improved in future commits and it will allow
the user to create and edit recipes.