feat: add backend integration
Now the recipes are stored and fetched from a backend, in this case
Firebase.

A new service, data-storage.service.ts, has been created to handle the
http requests to the backend.

To avoid errors when the user reload the app while seeing/editing a
recipe a resolver has been added, recipes-resolver.service.ts

This resolver fetchs the data when the user clicks/reloads in a recipe
or editing a recipe.

The commit contains all the work necessary to achieve the work described
above